core data

+++ FakeShop/Model/ProductResult.swift
@@ -9,7 +9,7 @@ import Foundation
 
 // MARK: - Product
 struct Products: Codable {
-    var id: Int = 0
+    var id: Int16 = 0
     var title: String = ""
     var price: Double = 0.0
     var description: String = ""
